修改了添加修改考试的界面功能，确保符合领导要求，再不加钱不干了

+++ scores/src/main/java/com/hysro/scores/service/IExamExcellentScoreLineService.java
@@ -34,7 +34,7 @@ public interface IExamExcellentScoreLineService
      * @param examExcellentScoreLine 优秀分数线
      * @return 结果
      */
-    public int insertExamExcellentScoreLine(ExamExcellentScoreLine examExcellentScoreLine);
+    public Long insertExamExcellentScoreLine(ExamExcellentScoreLine examExcellentScoreLine);
 
     /**
      * 修改优秀分数线
